feat: add support to double XPRV
We are moving the business logic related to the XPRV length to the witnet.dart
library (https://github.com/witnet/witnet.dart). So, in this PR, we only remove
the XPRV length validation while importing a wallet. The logic to handle a
double XPRV is implemented here: https://github.com/witnet/witnet.dart/pull/31