feat(wallet_sync): cache hash calls to the explorer
During the synchronization process, the wallet was calling multiple
times to the hash endpoint of the Witnet Block Explorer. It occurred
when the user had multiple accounts and had sent funds between them.
During the wallet synchronization, if two accounts of the same wallet
had sent funds between them, both would call the explorer with the
same hash of a valueTransferOutput. Because of that, the number of
repeated calls relied on the number of times accounts of the same
wallet have interacted between them.

To avoid multiple calls, we have implemented a "read through"
pattern to fetch the Witnet Block Explorer API only if the
queried value isn't stored in the database.